Minor changes in TransactionDetailView
