+++ CreditCardPayments/CreditCardPayments/Assets.xcassets/3.intro.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "3.intro3.png",
+      "filename" : "3.Intro3.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
